// frontend/vue/src/store/actions.type.ts
export enum ActionsType {
    SONG_PUBLISH = "publishSong",
    SONG_DELETE = "deleteSong",
    SONG_RESET_STATE = "resetSongState",
    CHECK_AUTH = "checkAuth",
    FAVORITE_ADD = "addFavorite",
    FAVORITE_REMOVE = "removeFavorite",
    FETCH_SONG = "fetchSong",
    FETCH_SONGS = "fetchSongs",
    FETCH_PROFILE = "fetchProfile",
    FETCH_PROFILE_FOLLOW = "fetchProfileFollow",
    FETCH_PROFILE_UNFOLLOW = "fetchProfileUnfollow",
    LOGIN = "login",
    LOGOUT = "logout",
    REGISTER = "register",
    UPDATE_USER = "updateUser"
}// end_ActionsTypes