// frontend/vue/src/store/mutations.type.ts
export enum MutationsType {
    FETCH_END = "setSongs",
    FETCH_START = "setLoading",
    PURGE_AUTH = "logOut",
    SET_SONG = "setSong",
    SET_AUTH = "setUser",
    SET_ERROR = "setError",
    SET_PROFILE = "setProfile",
    UPDATE_SONG_IN_LIST = "updateSongInList",
    RESET_STATE = "resetModuleState"
}// end_MutationsType